Fix to plink being killed (returncode=-9) during bmerge
We found plink command run getting killed while doing a bmerge of the inoput data and the 1K Genomes during the ancestry pruning step on the addi_qc_pipeline script. \n We understood there was a kill signal thrown to the plink command on the running environment, maybe due to a resource limits hit. \n To fix we filter out non common SNPs before the actual bmerge