Multi plot updates
Modifying script to plot results as well as channel to pass data onto the process. Now we take the full GWAS summary stats and avoid doing the merge across chromosomes within the process. In addition, the use of env variables, and the naming and quality of plots have been improved